+++ app/utils/WkConf.scala
@@ -12,6 +12,7 @@ class WkConf @Inject()(configuration: Configuration) extends ConfigReader {
   object Application {
 
     val insertInitialData = get[Boolean]("application.insertInitialData")
+    val insertLocalConnectDatastore = get[Boolean]("application.insertLocalConnectDatastore")
     val title = get[String]("application.title")
 
     object Authentication {
@@ -125,5 +126,16 @@ class WkConf @Inject()(configuration: Configuration) extends ConfigReader {
   }
 
   val children =
-    List(Application, Http, Mail, WebKnossos, Datastore, User, Braintracing, Features, Silhouette, Airbrake, Google)
+    List(Application,
+         Http,
+         Mail,
+         WebKnossos,
+         Datastore,
+         Tracingstore,
+         User,
+         Braintracing,
+         Features,
+         Silhouette,
+         Airbrake,
+         Google)
 }