+++ tools/proxy/proxy.js
@@ -52,6 +52,10 @@ const processes = {
       shell: true,
     },
   ),
+  connect: spawnIfNotSpecified("noConnect", "docker-compose run --rm webknossos-connect", [], {
+    cwd: ROOT,
+    shell: true,
+  }),
 };
 
 function spawnIfNotSpecified(keyword, command, args, options) {